--- frontend/helpers/project.ts
@@ -1,7 +1,0 @@
-export const PROJECT_CATEGORY_COLORS = {
-  'sustainable-agrosystems': '#E7C343',
-  'tourism-and-recreation': '#4492E5',
-  'forestry-and-agroforestry': '#E57D57',
-  'non-timber-forest-production': '#404B9A',
-  'human-capital-and-inclusion': '#A0616A',
-};
